requirements.txt
robotframework==5.0.1
pytest==7.1.2
python-dotenv==0.20.0